+++ src/Hl7.Fhir.Base.FhirPath.Validator/SymbolTable/SymbolTable.cs
@@ -64,6 +64,10 @@ namespace Hl7.Fhir.FhirPath.Validator
 			Add(new FunctionDefinition("answers", true, true) { GetReturnType = ReturnsAnswers, SupportsContext = AnswersSupportsContext }).AddValidation(ValidateNoArguments);
 			Add(new FunctionDefinition("ordinal").AddContexts(mi, "code-decimal,Coding-decimal"));
 			// sum, min, max, avg - shortcuts for aggregate -- need to check valid contexts, must be comparable, and incrementable
+			Add(new FunctionDefinition("sum", true).AddContexts(mi, "integer-integer,decimal-decimal,Quantity-Quantity")).AddValidation(ValidateNoArguments);
+			Add(new FunctionDefinition("min", true).AddContexts(mi, "integer-integer,decimal-decimal,Quantity-Quantity")).AddValidation(ValidateNoArguments);
+			Add(new FunctionDefinition("max", true).AddContexts(mi, "integer-integer,decimal-decimal,Quantity-Quantity")).AddValidation(ValidateNoArguments);
+			Add(new FunctionDefinition("avg", true).AddContexts(mi, "integer-integer,decimal-decimal,Quantity-Quantity")).AddValidation(ValidateNoArguments);
 		}
 
 		public FunctionDefinition Add(FunctionDefinition item)