+++ src/Test.FhirPath.Validator/BasicTests.cs
@@ -71,6 +71,22 @@ namespace Test.Fhir.FhirPath.Validator
 		}
 
         [TestMethod]
+		public void TestMethodSum()
+		{
+			string expression = "telecom.rank.sum()";
+			Console.WriteLine(expression);
+			var visitor = new FhirPathExpressionVisitor();
+			visitor.AddInputType(typeof(Patient));
+			var pe = _compiler.Parse(expression);
+			Console.WriteLine("---------");
+			var r = pe.Accept(visitor);
+			Console.WriteLine(visitor.ToString());
+			Console.WriteLine(visitor.Outcome.ToXml(new FhirXmlSerializationSettings() { Pretty = true }));
+			Assert.AreEqual("integer", r.ToString());
+			Assert.IsTrue(visitor.Outcome.Success);
+		}
+
+		[TestMethod]
         public void TestMethodWhere()
         {
             string expression = "contact.telecom.where(use='phone').system";