+++ changelog.md
@@ -10,6 +10,8 @@
    e.g. error display: `prop 'value' not found on ???` and a log entry: `.value : ???`
 * When a property is encountered that doesn't exist in the given context and that property name exists as a variable, 
    provide a message in the error that indicates that a variable of that name exists (help common typo issue)
+* Added SDC functions sum, min, max, avg
+* Added support for positiveInt/unsignedInt/base64Binary/date/instant type conversions for context processing to fhir primitive types
 
 ### November 20 2023: 5.3.0-beta5
 * Validate the `iif` function to be valid across any context (but not a collection)